Assertions in the STS rules.
We've had a couple of cases recently where we found bugs which resulted
from the ledger state getting into a position where it shouldn't be.
This commit starts the process of adding in assertions into the ledger
code to validate such conditions.

It also adds the ability to `reApply` as a bonus!